API

Contents

API

Contents

Supernovae Functions

Public Objects

Private Objects

Supernovae.RunModule Functions

Public Objects

Private Objects

Supernovae.RunModule.FilterModule Functions

Public Objects

Private Objects

Supernovae Functions

Public Objects

Supernovae.main — Method

main(toml_path::AbstractString, verbose::Bool, profile::Bool)

Loads toml_path, and sets up logging with verbosity based on verbose. Runs run_Supernovae and if profile is true, also profiles the code.

Arguments

toml_path::AbstractString: Path to toml input file.

verbose::Bool: Set verbosity of logging

profile::Bool: If true, profile run_Supernovae

Supernovae.main — Method

main()

Read the args, prepare the input TOML and run the actual package functionality. Runs main(toml_path, verbose, profile).

Private Objects

Supernovae.get_args — Method

get_args()

Helper function to get the ARGS passed to julia.

Supernovae.RunModule Functions

Public Objects

Supernovae.RunModule.run_Supernovae — Method

run_Supernovae(toml::Dict{String, Any})

Main entrance function for the package

Arguments

toml::Dict{String, Any}: Input toml file containing all options for the package

Private Objects

Supernovae.RunModule.FilterModule Functions

Public Objects

Supernovae.RunModule.FilterModule.Filter — Type

Filter

Photometric filter transmission curve.

Fields

facility::String: Name of the filter's facility

instrument::String: Name of the filter's instrument

passband::String: Name of the filter's passband

wavelength::Vector{Å}: Transmission curve wavelength

transmission::Vector{Float64}: Transmission curve transmission

Supernovae.RunModule.FilterModule.Filter — Method

Filter(facility::String, instrument::String, passband::String, filter_file::AbstractString)

Make Filter object from filter_file transmission curve.

Arguments

facility::String: Name of the filter's facility

instrument::String: Name of the filter's instrument

passband::String: Name of the filter's passband

filter_file::AbstractString: Path to transmission curve file. Assumed to be a comma delimited wavelength,transmission file.

Supernovae.RunModule.FilterModule.Filter — Method

Filter(facility::String, instrument::String, passband::String, config::Dict{String, Any})

Make Filter object from config options. config must include "FILTERPATH" => path/to/transmissioncurve. If this file exists, the transmission curve will be loaded via Filter(facility::String, instrument::String, passband::String, filter_file::AbstractString), otherwise attempt to create Filter via Filter(facility::String, instrument::String, passband::String, svo::PyCall.PyObject) and the SVO FPS database.

Arguments

facility::String: Name of the filter's facility

instrument::String: Name of the filter's instrument

passband::String: Name of the filter's passband

config::Dict{String, Any}: Options for creating a Filter.

Supernovae.RunModule.FilterModule.Filter — Method

Filter(facility::String, instrument::String, passband::String, svo::PyCall.PyObject)

Make Filter object from svo transmission curve.

Arguments

facility::String: Name of the filter's facility

instrument::String: Name of the filter's instrument

passband::String: Name of the filter's passband

svo::Pycall.PyObject: SVO transmission curve

Supernovae.RunModule.FilterModule.synthetic_flux — Method

synthetic_flux(filter::Filter, T::Unitful.Quantity{Float64})

Calculates the flux of a blackbody at temperature T, as observed with the filter

Arguments

filter::Filter: The Filter through which the blackbody is observed

T::Unitful.Quantity{Float64}: The temperature of the blackbody

Private Objects

Supernovae.RunModule.FilterModule.planck — Method

planck(T::Unitful.Quantity{Float64}, λ::Unitful.Quantity{Float64})

Planck's law: Calculates the specral radiance of a blackbody at temperature T, emitting at wavelength λ

Arguments

T::Unitful.Quantity{Float64}: Temperature of blackbody

λ::Unitful.Quantity{Float64}: Wavelength of blackbody

Supernovae.RunModule.FilterModule.save_filter — Method

save_filter(filter::Filter, filter_dir::AbstractString)

Save filter to directory filter_dir.

Arguments

filter::Filter: The Filter to save.

filter_dir::AbstractString: The directory to save filter to.

Supernovae.RunModule.FilterModule.svo — Method

svo(facility::String, instrument::String, passband::String)

Attempt to get filter transmission curve from SVO. Uses the python package astroquery via PyCall.

Arguments

facility::String: SVO name for the filter's facility

instrument::String: SVO name for the filter's instrument

passband::String: SVO name for the filter's passband